thêm phê duyệt

+++ demo_oracle_db/src/main/java/com/example/demo_oracle_db/service/userPermisson/impl/UserPermissionServiceImpl.java
@@ -1,20 +1,20 @@
 package com.example.demo_oracle_db.service.userPermisson.impl;
 
+import com.example.demo_oracle_db.entity.Function;
 import com.example.demo_oracle_db.entity.Permission;
-import com.example.demo_oracle_db.entity.RolePermission;
 import com.example.demo_oracle_db.exception.DodException;
-import com.example.demo_oracle_db.repository.FunctionRepository;
-import com.example.demo_oracle_db.repository.PermissionRepository;
-import com.example.demo_oracle_db.repository.RolePermissionRepository;
-import com.example.demo_oracle_db.repository.RoleRepository;
+import com.example.demo_oracle_db.repository.*;
+import com.example.demo_oracle_db.service.product.response.SearchBox;
 import com.example.demo_oracle_db.service.role.response.FunctionDto;
+import com.example.demo_oracle_db.service.role.response.PermissionDto;
 import com.example.demo_oracle_db.service.role.response.RolePermissionRes;
 import com.example.demo_oracle_db.service.userPermisson.UserPermissionService;
 import com.example.demo_oracle_db.service.userPermisson.request.*;
 import com.example.demo_oracle_db.util.MessageCode;
-import jakarta.persistence.criteria.JoinType;
-import jakarta.persistence.criteria.Predicate;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.criteria.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -30,18 +30,52 @@ public class UserPermissionServiceImpl implements UserPermissionService {
     @Autowired
     RolePermissionRepository rolePermissionRepository;
     @Autowired
+    FunctionRoleRepository functionRoleRepository;
+    @Autowired
     RoleRepository roleRepository;
+    @Autowired
+    EntityManager entityManager;
 
     @Override
     public RolePermissionRes getPermissionsByRole(Long roleId) throws DodException {
         String roleName = roleRepository.findNameById(roleId).orElseThrow(()->new DodException(MessageCode.ROLE_NOT_FOUND));
-        List<Permission> permissions = permissionRepository.findPermissionsByRole(List.of(roleId));
-        List<FunctionDto> functions = FunctionDto.mapByPermission(permissions);
+        List<FunctionDto> functions = getFunctionsByRoleId(roleId);
         RolePermissionRes res = new RolePermissionRes();
         res.setId(roleId);
-        res.setName(roleName);
+        res.setName(roleName.replace("ROLE_", "").replace("_", " "));
         res.setFunctions(functions);
         return res;
+    }
+
+    private List<FunctionDto> getFunctionsByRoleId(Long roleId) {
+        CriteriaBuilder cb = entityManager.getCriteriaBuilder();
+        CriteriaQuery<FunctionDto> query = cb.createQuery(FunctionDto.class);
+        Root<Function> root = query.from(Function.class);
+        query.multiselect(root.get("id"), root.get("name"), root.get("feRoute"));
+        List<Predicate> predicates = new ArrayList<>();
+        predicates.add(cb.equal(root.join("functionRoleList").get("roleId"), roleId));
+        query.where(cb.and(predicates.toArray(new Predicate[0])));
+
+        List<FunctionDto> functions = entityManager.createQuery(query).getResultList();
+        for (FunctionDto function: functions
+             ) {
+            function.setPermissions(getPermissionsByRoleId(roleId, function.getId()));
+        }
+        return functions;
+    }
+
+    private List<PermissionDto> getPermissionsByRoleId(Long roleId, Long functionId) {
+        CriteriaBuilder cb = entityManager.getCriteriaBuilder();
+        CriteriaQuery<PermissionDto> query = cb.createQuery(PermissionDto.class);
+        Root<Permission> root = query.from(Permission.class);
+        query.multiselect(root.get("id"), root.get("name"), root.get("beEndPoint"), root.get("defaultPermission"));
+        List<Predicate> predicates = new ArrayList<>();
+        predicates.add(cb.and(
+                cb.equal(root.get("functionId"), functionId),
+                cb.equal(root.join("rolePermissionList").get("roleId"), roleId))
+        );
+        query.where(cb.and(predicates.toArray(new Predicate[0])));
+        return entityManager.createQuery(query).getResultList();
     }
 
     @Override
@@ -54,40 +88,26 @@ public class UserPermissionServiceImpl implements UserPermissionService {
         if (!functionRepository.existsById(request.getFunctionId())) {
             throw new DodException(MessageCode.FUNCTION_NOT_FOUND);
         }
-
-        if (rolePermissionRepository.exists((root, query, criteriaBuilder) -> {
-            List<Predicate> predicates = new ArrayList<>();
-            predicates.add(criteriaBuilder.equal(root.get("roleId"), request.getRoleId()));
-            predicates.add(criteriaBuilder.equal(root.join("permission", JoinType.INNER).get("functionId"), request.getFunctionId()));
-            return criteriaBuilder.and(predicates.toArray(predicates.toArray(new Predicate[0])));
-        })) {
+        if(functionRoleRepository.existsByFunctionIdAndRoleId(request.getFunctionId(), request.getRoleId())){
             throw new DodException(MessageCode.FUNCTION_ALREADY_ADD);
         }
 
-        List<Permission> permissions = permissionRepository.findAll((root, query, criteriaBuilder) -> {
-            List<Predicate> predicates = new ArrayList<>();
-            predicates.add(criteriaBuilder.equal(root.get("functionId"), request.getFunctionId()));
-            predicates.add(criteriaBuilder.equal(root.get("defaultPermission"), 1));
-            return criteriaBuilder.and(predicates.toArray(new Predicate[0]));
-        });
-        List<RolePermission> rolePermissionList = new ArrayList<>();
-        for (Permission permission : permissions
+        functionRoleRepository.addFunctionRole(request.getFunctionId(), request.getRoleId());
+
+        List<Long> permissions = permissionRepository.getIdByFunctionIdAndDefaultPermission(request.getFunctionId(), 1);
+        for (Long permission : permissions
         ) {
-            RolePermission rolePermission = new RolePermission();
-            rolePermission.setPermissionId(permission.getId());
-            rolePermission.setRoleId(request.getRoleId());
-            rolePermissionList.add(rolePermission);
+            rolePermissionRepository.addPermissionRole(permission, request.getRoleId());
         }
-        rolePermissionRepository.saveAll(rolePermissionList);
     }
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void addUserNewFunction(AddUserNewFunction request) throws DodException {
         if (!roleRepository.existsById(request.getRoleId())) {
-            throw new DodException(MessageCode.ROLE_NOT_FOUND);
+            throw new DodException(MessageCode.ROLE_NOT_FOUND, HttpStatus.NOT_FOUND);
         }
-        if (functionRepository.existsByFunctionName(request.getFunctionName())) {
+        if (functionRepository.existsByName(request.getFunctionName())) {
             throw new DodException(MessageCode.FUNCTION_NAME_EXISTS, request.getFunctionName());
         }
         if (functionRepository.existsByFeRoute(request.getFeRoute())) {
@@ -114,7 +134,33 @@ public class UserPermissionServiceImpl implements UserPermissionService {
         if (!functionRepository.existsById(request.getFunctionId())) {
             throw new DodException(MessageCode.FUNCTION_NOT_FOUND);
         }
+        if(!functionRoleRepository.existsByFunctionIdAndRoleId(request.getFunctionId(), request.getRoleId())){
+            throw new DodException(MessageCode.FUNCTION_ROLE_NOT_FOUND);
+        }
+        functionRoleRepository.deleteByFunctionIdAndRoleId(request.getFunctionId(),request.getRoleId());
         rolePermissionRepository.deleteByFunctionId(request.getFunctionId(), request.getRoleId());
+    }
+
+    @Override
+    public List<SearchBox> getFunctionSearch(Long id) throws DodException {
+        if(!roleRepository.existsById(id)) throw new DodException(MessageCode.ROLE_NOT_FOUND);
+        return functionRepository.getFunctionRoleNot(id).stream().map(SearchBox::new).toList();
+    }
+
+    @Override
+    public List<SearchBox> getPermissionSearch(Long roleId, Long functionId) throws DodException {
+        if(!roleRepository.existsById(roleId)) throw new DodException(MessageCode.ROLE_NOT_FOUND);
+        return permissionRepository.getPermissionRoleNot(roleId, functionId).stream().map(SearchBox::new).toList();
+    }
+
+    @Override
+    public List<PermissionDto> getUserFunctionPermission(Long roleId, Long functionId) throws DodException {
+        if(!roleRepository.existsById(roleId)) throw new DodException(MessageCode.ROLE_NOT_FOUND);
+        if(!functionRoleRepository.existsByFunctionIdAndRoleId(functionId, roleId)){
+            throw new DodException(MessageCode.FUNCTION_ROLE_NOT_FOUND);
+        }
+        List<Permission> permissions = permissionRepository.getUserFunctionPermission(roleId, functionId);
+        return permissions.stream().map(PermissionDto::new).toList();
     }
 
     @Override
@@ -125,6 +171,10 @@ public class UserPermissionServiceImpl implements UserPermissionService {
         }
         if (!permissionRepository.existsById(request.getPermissionId())) {
             throw new DodException(MessageCode.PERMISSION_NOT_FOUND);
+        }
+        Long functionId = permissionRepository.getFunctionIdById(request.getPermissionId());
+        if(!functionRoleRepository.existsByFunctionIdAndRoleId(functionId, request.getRoleId())){
+            throw new DodException(MessageCode.FUNCTION_ROLE_NOT_FOUND);
         }
         if (rolePermissionRepository.existsByRoleIdAndPermissionId(request.getRoleId(), request.getPermissionId())) {
             throw new DodException(MessageCode.PERMISSION_ALREADY_ADD);
@@ -158,11 +208,14 @@ public class UserPermissionServiceImpl implements UserPermissionService {
         if (!roleRepository.existsById(request.getRoleId())) {
             throw new DodException(MessageCode.ROLE_NOT_FOUND);
         }
-        if (!functionRepository.existsById(request.getPermissionId())) {
+        if (!permissionRepository.existsById(request.getPermissionId())) {
             throw new DodException(MessageCode.PERMISSION_NOT_FOUND);
         }
         if(!rolePermissionRepository.existsByRoleIdAndPermissionId(request.getRoleId(), request.getPermissionId())){
             throw new DodException(MessageCode.ROLE_PERMISSION_NOT_FOUND);
+        }
+        if(permissionRepository.existsByIdAndDefaultPermission(request.getPermissionId(), 1)){
+            throw new DodException(MessageCode.PERMISSION_DEFAULT);
         }
         rolePermissionRepository.deleteByRoleIdAndPermissionId(request.getRoleId(),request.getPermissionId());
     }
